--- a/Cargo.toml
+++ b/Cargo.toml
@@ -14,10 +14,9 @@ keywords = ["regex", "text", "generation"]
 categories = ["text-processing"]
 
 [dependencies]
-rand = "0.3.15"
-regex-syntax = "0.4.1"
-byteorder = "1.0.0"
-error-chain = "0.10.0"
+rand = "0.5.5"
+regex-syntax = "0.6.2"
+error-chain = "0.12.0"
 
 [dev-dependencies]
-regex = "0.2.2"
+regex = "1.0.0"